requirements.txt
# Connexion Odoo
OdooRPC==0.10.1
requests==2.32.3

# Base de données
psycopg==3.2.3
sqlalchemy==2.0.23
supabase==2.0.4

# Utils
python-dotenv==1.0.0
pandas==2.2.3
python-dateutil==2.8.2
numpy==2.0.2